Rewrite the inline WebKit view in SwiftUI
